+++ changelog.md
@@ -277,3 +277,4 @@
 * Updated translations
 * Fixed crash on dedicated server when picking up children
 * Cleaned up config, linked to config wiki
+* Villager can no longer plant modded plants to remove a crash