+++ changelog.md
@@ -274,8 +274,9 @@
 * Fixed young villagers not holding tools correctly
 
 # 7.3.0
+
 * Updated translations
 * Fixed crash on dedicated server when picking up children
-* Cleaned up config, linked to config wiki
+* Cleaned up config, added link to config wiki
 * Villager can no longer plant modded plants to remove a crash
 * Fixed a few crashes