+++ changelog.md
@@ -280,3 +280,7 @@
 * Cleaned up config, added link to config wiki
 * Villager can no longer plant modded plants to remove a crash
 * Fixed a few crashes
+
+# 7.3.0
+
+* Fixed advancements and book rewards